requirements.txt
ollama
pyperclip
python-dotenv